generator/requirements.txt
kafka-python
sanic
sanic-redis
aiopg[sa]